requirements.txt
matplotlib==3.9.2
pygame
#numpy==2.2.5
numpy
pandas==2.2.3
pillow==11.0.0
#psycopg2==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.1
scipy==1.13.1
imbalanced-learn
seaborn==0.13.2
pipreqs==0.5.0
rasterio==1.4.1
aiofiles==24.1.0

torch==2.7.0
torchaudio==2.7.0
torchinfo==1.8.0
torchmetrics==1.7.1
torchvision==0.22.0

wfdb
sentencepiece
transformers  #richiede sentencepiece
datasets

pytorch-lightning==2.5.1.post0
lightning==2.5.1.post0
lightning-utilities==0.14.3

tensorboard==2.19.0
colorama==0.4.6
tqdm==4.67.1